@@ -4,16 +4,16 @@ apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion "27.0.3"
+    compileSdkVersion 28
+    buildToolsVersion "28.0.3"
 
     defaultConfig {
         applicationId "fr.corenting.edcompanion"
         minSdkVersion 19
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 27
         versionName "2.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         vectorDrawables.useSupportLibrary = true
 
         buildConfigField "String", "INARA_API_KEY", '"' + INARA_API_KEY + '"'
@@ -50,19 +50,20 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
-    androidTestImplementation('com.android.support.test.espresso:espresso-core:2.2.2', {
+    androidTestImplementation('androidx.test.espresso:espresso-core:3.1.0-alpha4', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
 
-    // Support libs
-    ext.supportLibsVersion = "27.1.1"
-    implementation "com.android.support:support-v4:$supportLibsVersion"
-    implementation "com.android.support:support-compat:$supportLibsVersion"
-    implementation "com.android.support:preference-v7:$supportLibsVersion"
-    implementation "com.android.support:preference-v14:$supportLibsVersion"
-    implementation "com.android.support:design:$supportLibsVersion"
-    implementation "com.android.support:cardview-v7:$supportLibsVersion"
-    implementation "com.android.support:support-vector-drawable:$supportLibsVersion"
+    // AndroidX
+    ext.androidXVersion = "1.0.0"
+    implementation "androidx.legacy:legacy-support-v4:$androidXVersion"
+    implementation "androidx.core:core:$androidXVersion"
+    implementation "androidx.core:core:$androidXVersion"
+    implementation "androidx.preference:preference:$androidXVersion"
+    implementation "androidx.legacy:legacy-preference-v14:$androidXVersion"
+    implementation "com.google.android.material:material:$androidXVersion"
+    implementation "androidx.cardview:cardview:$androidXVersion"
+    implementation "androidx.vectordrawable:vectordrawable:$androidXVersion"
 
     // Play services / Firebase
     implementation 'com.google.firebase:firebase-messaging:17.3.2'
@@ -74,14 +75,14 @@ dependencies {
     implementation "com.squareup.retrofit2:converter-gson:$retrofitVersion"
 
     // Butterknife
-    ext.butterKnifeVersion = "8.8.1"
+    ext.butterKnifeVersion = "9.0.0-SNAPSHOT"
     implementation "com.jakewharton:butterknife:$butterKnifeVersion"
     annotationProcessor "com.jakewharton:butterknife-compiler:$butterKnifeVersion"
 
     // Other
     implementation 'com.google.code.gson:gson:2.8.5'
     implementation 'com.jakewharton.threetenabp:threetenabp:1.1.0'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.0-alpha2'
     implementation 'org.greenrobot:eventbus:3.1.1'
     implementation 'me.zhanghai.android.materialprogressbar:library:1.4.2'
     implementation "com.mikepenz:itemanimators:1.0.2"
