@@ -1,9 +1,9 @@
 package fr.corenting.edcompanion.adapters;
 
 import android.content.Context;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentPagerAdapter;
 
 import fr.corenting.edcompanion.R;
 import fr.corenting.edcompanion.fragments.SystemDetailsFragment;
