@@ -1,7 +1,7 @@
 package fr.corenting.edcompanion.network;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import org.greenrobot.eventbus.EventBus;
 
