@@ -1,8 +1,8 @@
 package fr.corenting.edcompanion.utils;
 
 import android.content.Context;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
 import android.util.Log;
 import android.util.TypedValue;
 import android.view.View;
